chore: move cache client and cache client builder under cache mod
Moves the cache client and cache client builder under the cache
mod.

We keep the Cache Client struct exported at the library level.

We also add the Topic Client struct exported at the library level.
