+++ src/cache/mod.rs
@@ -99,5 +99,8 @@ pub use config::configuration::Configuration;
 pub use config::configurations;
 
 mod collection_ttl;
-
 pub use collection_ttl::CollectionTtl;
+
+mod cache_client;
+mod cache_client_builder;
+pub use cache_client::CacheClient;
