+++ src/cache/requests/scalar/set.rs
@@ -1,6 +1,6 @@
 use crate::cache::requests::MomentoRequest;
-use crate::cache_client::CacheClient;
 use crate::utils::prep_request_with_timeout;
+use crate::CacheClient;
 use crate::{IntoBytes, MomentoResult};
 use std::time::Duration;
