+++ src/lib.rs
@@ -1,18 +1,19 @@
 pub mod cache;
+pub use cache::CacheClient;
+
 pub mod config;
-pub mod errors;
-pub mod topics;
-
-mod cache_client;
-mod cache_client_builder;
 mod credential_provider;
-mod grpc;
-mod utils;
+pub use credential_provider::CredentialProvider;
 
-pub use self::errors::*;
-pub use crate::cache_client::CacheClient;
-pub use crate::credential_provider::CredentialProvider;
+pub mod errors;
+pub use errors::*;
+
+mod grpc;
+
+pub mod topics;
+pub use topics::TopicClient;
+
+mod utils;
+pub use crate::utils::{IntoBytes, IntoBytesIterable};
 
 pub type MomentoResult<T> = Result<T, MomentoError>;
-
-pub use crate::utils::{IntoBytes, IntoBytesIterable};
